+++ src/cognitect/aws/client/api.clj
@@ -26,16 +26,16 @@
 
   :api                  - required, name of the api you want to interact with e.g. s3, cloudformation, etc
   :http-client          - optional, to share http-clients across aws-clients
-                          Default: default-http-client
+                          Default: cognitect.aws.client.shared/http-client
   :region-provider      - optional, implementation of aws-clojure.region/RegionProvider
-                          protocol, defaults to cognitect.aws.region/default-region-provider.
+                          protocol, defaults to cognitect.aws.client.shared/region-provider.
                           Ignored if :region is also provided
   :region               - optional, the aws region serving the API endpoints you
                           want to interact with, defaults to region provided by
-                          by the region-provider
+                          the region-provider
   :credentials-provider - optional, implementation of
                           cognitect.aws.credentials/CredentialsProvider protocol
-                          Default: cognitect.aws.credentials/default-credentials-provider
+                          Default: cognitect.aws.client.shared/credentials-provider
   :endpoint-override    - optional, map to override parts of the endpoint. Supported keys:
                             :protocol     - :http or :https
                             :hostname     - string
@@ -81,7 +81,7 @@
         endpoint-provider    (endpoint/default-endpoint-provider
                               (get-in service [:metadata :endpointPrefix])
                               endpoint-override)]
-    (dynaload/load-ns (symbol (str "cognitect.aws.protocols." (get-in service [:metadata :protocol]))))
+    (dynaload/load-ns (symbol (str "cognitect.aws.protocols." (service/service-protocol service))))
     (client/->Client
      (atom {'clojure.core.protocols/datafy (fn [c]
                                              (let [info (client.protocol/-get-info c)
@@ -104,7 +104,8 @@
       :validate-requests?   (atom nil)})))
 
 (defn default-http-client
-  "Create an http-client to share across multiple aws-api clients."
+  "Returns a new instance of the default type of http client. This function may be used to create a
+  single http-client instance to share across multiple aws-api clients."
   []
   (http/resolve-http-client nil))